Absorbs: fix stacked and disable lookAhead
-"Stacked" absorb style was overflowing as well, so I implemented dedicated logic to handle the cases with incoming heal and overAbsorb.

LookAhead was misbehaving with erratic heal prediction on HoTs (my case was Rejuvenation for instance would predict correctly the first tick, but if target frame was present, it would predict more heal for a brief moment, then jump to next tick prediction, instead of predicting next tick). This was also affecting overflowing due to incomingHeal argument that was being passed into the updater would not match the prediction.