Heal Prediction: fix lookAhead being ignored on multiple frames
Related to #11.
https://github.com/Zidras/ElvUI_ProjectZidras/commit/6310e2a56f2187033a03a4e678b4791bef56d76c

If there was more than one unitframe with the same GUID, the Heal Prediction bar would always assume no lookAhead on the previous unitframes. After some investigation, I concluded that HealComm4 element from ElvUI was somehow conflicting with PZ, and since it is only being used for ElvUI HP, it is best to disable it completely and solely rely on PZ element.

Another alternative that gave satisfying results would be defining self.HealCommTimeframe, although it would be wasting cpu for no reason.